fix: write edge weights as float to ensure DOUBLE type for neug leiden
neug COPY TEMP infers INT64 for integer CSV values (e.g. '2'), but
weighted leiden produces degenerate results (every node its own
community) with INT64 weights. Writing as float ('2.0') ensures
DOUBLE inference and leiden works correctly.

Verified on neug codebase (1332 files, 2982 edges):
- INT64 weights: 1332 communities (1 per file) at any resolution
- DOUBLE weights: 312 communities with reasonable size distribution

+++ graphify/storage.py
@@ -636,7 +636,7 @@ def _aggregate_file_edges(conn: object, csv_path: Path) -> set[str]:
         writer = csv.writer(f)
         writer.writerow(["from_file", "to_file", "weight"])
         for from_file, to_file, weight in results:
-            writer.writerow([from_file, to_file, int(weight)])
+            writer.writerow([from_file, to_file, float(weight)])
             all_files.add(from_file)
             all_files.add(to_file)
